requirements.txt
transformers==4.57.0
accelerate==1.10.1
torch==2.7.1+cu118
torchvision0.22.1+cu118
pillow==11.3.0
numpy==2.3.3
matplotlib==3.10.7
scipy==1.16.2